feat(api): modify CRUD Room

+++ agos-api/src/main/java/UFC/Agos/services/imp/RoomService.java
@@ -3,6 +3,7 @@ package UFC.Agos.services.imp;
 import UFC.Agos.models.Department;
 import UFC.Agos.models.Room;
 import UFC.Agos.repositories.RoomRepository;
+import UFC.Agos.repositories.ThesisRepository;
 import UFC.Agos.services.IRoomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +15,8 @@ import java.util.Objects;
 public class RoomService implements IRoomService {
     @Autowired
     RoomRepository roomRepository;
+    @Autowired
+    ThesisRepository thesisRepository;
     @Override
     public List<Room> getRooms() {
         return roomRepository.findAll();
@@ -32,8 +35,12 @@ public class RoomService implements IRoomService {
     @Override
     public void deleteRoom(Long roomId) throws Exception {
         boolean exists = roomRepository.existsById(roomId);
+        boolean thesisExist = thesisRepository.getThesesByRoom(roomId).isEmpty();
         if(!exists){
             throw new IllegalStateException("The room with id "+ roomId + " does not exist");
+        }
+        if(!thesisExist){
+            throw new IllegalStateException("the room with id "+ roomId +" can't be removed because it contains thesis");
         }
         //should add something here !!!
         roomRepository.deleteById(roomId);
